requirements.txt
schema~=0.7.2
setuptools~=50.3.2
PyYAML~=5.3.1
parameterized~=0.7.4